pipeline/requirements.txt
# Core dependencies
pandas>=2.0.0
requests>=2.31.0

# Error handling and retry logic
tenacity>=8.2.0

# Configuration management
pyyaml>=6.0

# Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking
mypy>=1.5.0

# Code quality (optional but recommended)
black>=23.7.0
flake8>=6.1.0

# Progress bars (optional)
tqdm>=4.66.0

# Gmail API dependencies
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0